requirements.txt
Django==1.6.2
Pillow==2.3.0
South==0.8.4
Unipath==1.0
boto==2.25.0
dj-database-url==0.2.2
dj-static==0.0.5
django-boto==0.3.0
django-cors-headers==0.12
django-rest-swagger==0.1.11
django-toolbelt==0.0.1
djangorestframework==2.3.12
gunicorn==18.0
pillowfight==0.2
psycopg2==2.5.2
pystache==0.5.3
requests==2.2.1
selenium==2.40.0
static==1.0.2
urbanairship==0.6.2
wsgiref==0.1.2
